Move common internal API usage detection to parent class
